requirements.txt
torch==2.0.1
openai_whisper==20231117
pydub==0.25.1
